+++ src/main/kotlin/dev/ekvedaras/laravelquery/utils/ModelRelationResolver.kt
@@ -10,6 +10,7 @@ import com.jetbrains.php.lang.psi.elements.impl.ParameterListImpl
 import com.jetbrains.php.lang.psi.elements.impl.PhpReturnImpl
 import com.jetbrains.php.lang.psi.elements.impl.StringLiteralExpressionImpl
 import dev.ekvedaras.laravelquery.models.DbReferenceExpression
+import dev.ekvedaras.laravelquery.utils.ClassUtils.Companion.isChildOf
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isInsideRelationClosure
 import dev.ekvedaras.laravelquery.utils.PsiUtils.Companion.unquoteAndCleanup
 
@@ -21,19 +22,36 @@ class ModelRelationResolver(
         val modelReference: PhpTypedElement = tableAndAliasCollector.resolveModelReference(methods) ?: return
         val model = modelReference.getClass(reference.project) ?: return
 
+        // $user->customer()->create(['']). Detect customer() and resolve table name as customers table only
+        val relationMethod = methods.lastOrNull { mm ->
+            MethodUtils.resolveMethodTypeClasses(mm, method.project).any {
+                it.isChildOf(LaravelClasses.Relation)
+            }
+        }
+        if (relationMethod != null && relationMethod.name != null) {
+            resolveRelationTable(relationMethod.name ?: "", model)
+            return
+        }
+
         tableAndAliasCollector.resolveTableName(model)
 
         val deepParent =
-            method.parent?.parent?.parent?.parent?.parent?.parent ?: return // TODO utilize parentOfType<>() ?
-
-        if (deepParent.isInsideRelationClosure()) {
+            method.parent?.parent?.parent?.parent?.parent?.parent // TODO utilize parentOfType<>() ?
+        if (deepParent?.isInsideRelationClosure() == true) {
             resolveRelationTable(deepParent, model)
         }
     }
 
     @Suppress("ReturnCount")
     private fun resolveRelationTable(deepParent: PsiElement, model: PhpClass) {
-        val relationName = deepParent.firstChild.text.unquoteAndCleanup()
+        resolveRelationTable(
+            deepParent.firstChild.text.unquoteAndCleanup(),
+            model
+        )
+    }
+
+    @Suppress("ReturnCount")
+    private fun resolveRelationTable(relationName: String, model: PhpClass) {
         val relationMethod = model.methods.firstOrNull { it.name == relationName } ?: return
         if (relationMethod.lastChild !is GroupStatementImpl) {
             return